Deepfake Audio-Video Detection/requirements.txt
torch
torchvision
pandas
opencv-python
numpy
Pillow
timm
moviepy
librosa